requirements.txt
# Core per PySpark e Data Handling
pyspark==4.0.0
pandas==2.3.0
numpy==2.2.6
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
python-dotenv==1.1.0

# Visualizzazione per i Notebooks
matplotlib==3.10.3
seaborn==0.13.2
scikit-learn==1.7.0

# API per il download dei dati
kaggle==1.7.4.5

# Dipendenze per Jupyter Notebooks
jupyterlab==4.4.2
ipykernel==6.29.5